+++ junos_ops/__init__.py
@@ -17,4 +17,4 @@ See Also:
     https://github.com/shigechika/junos-ops
 """
 
-__version__ = "0.6.2"
+__version__ = "0.7.0"